source/requirements.txt
amqp==5.0.9
asgiref==3.4.1
asyncio==3.4.3
atomicwrites==1.4.0
attrs==21.4.0
authentication==1.1.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3
decorator==5.1.0
Deprecated==1.2.13
Django==3.2.8
django-environ==0.8.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
ffmpeg==1.4
Flask==2.0.2
future==0.18.2
geocoder==1.38.1
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
passlib==1.7.4
pika==1.2.0
pluggy==1.0.0
prompt-toolkit==3.0.26
py==1.11.0
pycryptodome==3.14.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
pytz==2021.3
PyYAML==6.0
ratelim==0.1.6
redis==4.1.2
requests==2.26.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.1
zipp==3.7.0
django-environ==0.8.1